VA/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai-whisper==20231117
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
pyttsx3>=2.90
trafilatura>=2.0.0
requests>=2.25.0
lxml_html_clean>=0.4.0
